Add Rust binary
Skeleton Rust binary with JSON processing from serde, and build steps
for running "cargo build" and copying the binary into the OSX app.

The build could be set up in a more sophisticated way (using Xcode's
designated build products directory rather than just writing target/
in the tree, following the release/debug build configuration, letting
Xcode be aware of the file dependencies), but the point of this commit
is to get something simple that works.