Working on generic rope implementation
Trying to abstract out details about what's being stored (the leaf
contents) what information is being gathered (node info) and what
metrics are used.

Still very much work in progress, but the signatures seem close to
something workable. It includes an interesting example of a metric,
an compact representation of a set of line breaks.