Adding an Interval abstraction
The fiddling with adding a "fudge" to the end of intervals when they
could include fragments felt really messy, so I'm replacing all that
with a proper Interval struct. This will also come in very handy when
we implement interval trees.

This commit also fixes a number of bugs in the old implementation,
offering something of a demonstration of how easy it was to make
off-by-one errors.