Move leaf slicing to Interval abstraction
The case analysis for fragments at the ends of leaves was getting
totally out of hand, so this patch switches it to using intervals,
representing that as whether the interval is open or closed at the
end.

Not 100% tested, but my confidence is a lot higher, and a bunch of
bugs from the previous revision are fixed.