Starting line wrapping
This commit is a big step towards adding line wrapping functionality.
It adds an optional "breaks" field to editor's view, which stores the
result of line breaking, and applies that while rendering.

It also computes the line breaks (in batch mode) on pressing the F1
key. Bad things will probably happen when editing the buffer after
that, and there are still some problems with cursor movement, but for
just scrolling it seems to work well.