Start line break iterator
There was no suitable implementation of the UAX #14 line break
iterator, so I started my own. It does property lookup with an
efficient trie, and a state machine computing the boundaries. Most of
the rules seem to be implemented aside from South East Asian
languages, but it hasn't been extensively tested yet.

Also, I haven't plumbed it to the word wrap yet, which was the
motivation for doing this.