More line break fixes, randomized testing
Added a utility to generate random test data using ICU, and test the
line breaker against it, as well as the official test suite. The
handling of SA is not quite right (SA combining marks are still
treated as AL), but other than that it seems to follow the spec pretty
well. The major discrepancy from ICU results is that ICU implements
the full regex in Example 7 of "Examples of Customization," where this
library implements LB25 literaly.

So I am now reasonably confident that the library follows the spec.
The one exception is the fallback being insensitive to general
category in the SA class.