Beginning implementation of spans
We want to store spans to annotate the text with color and font
changes.  This commit begins the implementation of those, using an
interval tree. At this point, much of the code is in place to
represent spans, but is not tested, and iteration through "next span
transition" still needs to be done.