More robust caching of lines in EditView
EditView now splits off the fetching of lines to a getLine() method
rather than having complex logic in drawRect(), and now is explicit
about storing lines as a cache. This cache is invalidated by a settext
message from the core, and is size-limited, but persists, so scrolling
up and down through a document that fits in the window (currently
hard-coded as 1000 lines) avoids RPC altogether.

Having a separate getLine() will also be useful for other methods
(mouse handling especially) that need to access line data.