Fix leaf-boundary logic in Cursor::prev
When the original cursor was at the start of a leaf, the prev() method
would acecpt the boundary at the end of the previous leaf.  This never
happens for Breaks (the start of a leaf is not considered a boundary),
which is how I missed it, but certainly does happen for string ropes.

Fixes #47. Thanks to @marchelzo for the report.