Don't force boundary at end of string
Allow `is_boundary` to return false if the leaf `is_boundary` method
indicates there isn't one there. This will happen if the last line
is partial (there's no terminating newline), so this change makes it
possible for clients to tell whether that's the case or not.

Useful for getting issue #59 right.