Add engine for updates and undo
Add a mini CRDT engine, which takes edits (potentially async, can be
based on earlier revisions) and undo groups, producing Delta objects
and snapshots of any revision in the history.

Untested, and lacking a method for gc of revisions that have been
fully committed, but the ideas should be quite clear.