Fix transforms in Engine::mk_new_rev
The transforms in mk_new_rev had subtle problems which caused
incorrect behavior when the edit contains both insertions and
deletions (as is the case when pasting when text is selected).

Also fixes an issue that would only show up when applying edits to
revisions other than head (which will happen when we get async
plugins).

Fixes #75.