Wire up plugin to access editor state
A bunch more development of the plugin interface, so that plugins
can fetch lines from the editor, and also set spans.

Also includes a more developed RPC peer on the Python plugin
prototype.

As a demo, the plugin highlights all occurrences of the word "fizz" in
red. However, it doesn't propagate the update to the front end, so the
spans don't show until the next update (say, on cursor movement).
There are lots of limitations and performance issues.