Create separate crate for RPC
The RPC peer is more aggressively threaded, making it suitable for
plugins as well as the core. There's also support for sending
synchronous RPC's, but this is not yet tested.

In addition, there's a method to detect whether there are any pending
requests. This is intended to be polled by plugins that do bulk work
that may take longer than a reasonable interactive response time.