Make RPC infrastructure not require Send
A bit of refactoring to run the handler on the main thread rather than
spinning up a handler thread. As a consequence, the handler function
doesn't have to be Send (and various bits of syntect state weren't).

It does spin up an IO thread, and we use a bit of callback trickery so
the reader itself doesn't have to move across a thread boundary
either.