Do highlighting as idle task
The plugin is now sensitive to other requests coming in, running
the highlighting as an idle task that can be interrupted.

A considerable amount of RPC refactoring was needed to make this work,
including making RPC handling a trait rather than an FnMut so that we
can pass a reference to the idle proc.

Overall performance is still not great, but this will get better when
we aggregate more than one line at a time from the plugin to the core.