Send text in updates to plugin
Updates to the plugin contain a delta indicating what's changed. This
patch includes the inserted text as well (as long as it fits in the
maximum RPC size).

The caching_plugin layer updates its own internal cache with this
text.  As a result, in many cases the plugin no longer has to query
the core for the text contents.