Send syntax highlighting in bulk
This patch makes the plugin send syntax highlighting results in larger
chunks than one line at a time. This is considerably more efficient,
especially as the core is not (yet) doing minimal invalidation to the
front-end, so each update also propagates to the front-end.

Also fixes potential crasher (selection can become invalid on view
update).

Still work in progress, as it doesn't apply operational transformation
to the spans.

Progress towards #23.