Add cmake dependency
I couldn't compile on OS X without install cmake. This pull request adds a step to install that.

I'm assuming homebrew is already installed, but that might of course not be the case. Just didn't know how verbose to make it. Feel free to ditch this pull request in favor of something more precise.