Reorganization of core into a library
Moving the core into a library, a step towards separating out the
json-rpc from the inner workings. One other change is the the type
of the RPC writer is now parametrized.

The protocol is still organized as json sent over byte-stream
channels, but it's no longer bound specifically to stdin/stdout.