Start PEG-based language plugin
This is the start of a new syntax coloring plugin that uses an
explicit state stack and a parser expression grammar approach.

The PEG is notable in that it doesn't rely on macros, just Rust
language features and fairly simple types. Even so, it is reasonably
concise and pleasant to use.