Send blank line (with cursor) on new tab
Schedule a render in the core's idle task on synchronous RPC (which,
at the moment, basically only a new tab). This send the actual render
of a new tab (a single blank line with a cursor) to the front-end.

Having an idle task opens the door for more asynchronous processing in
the future (such as line wrapping). Also, we could use the idle task
to batch up a single render for multiple edit operations in a row, but
it's not clear that will actually help much.

Fixes #161