Implemented New Tab menu item & ⌘+T
This uses approximately the same trick used in c74a522 to reuse empty
views, except in this case the global state is hidden as a property of
the Document type, as opposed to the AppDelegate.

If at some point in the future we write a document management model from
scratch, this can all become much cleaner and we can pass creation
params along with our new & open methods. For the time being, I think
this is an acceptable approach.

As of this change, we require macOS 10.12, to take advantage of the new
tabbing APIs described in this document: https://developer.apple.com/library/content/releasenotes/AppKit/RN-AppKit/#10_12Window%20Tabbing