Speculatively send lines to front-end on scroll
Before, rendering lines for scrolling was entirely reactive, and the
core wouldn't send lines until they were explicitly requested (which
often meant that a blank line had been displayed). With this patch,
we keep track of which lines the front-end has in its cache, and send
lines in the scroll region (plus slop) that it doesn't already have.

This is a step toward purely minimal invalidation, but that will
require even more accurate tracking of deltas.

Improves implementation of new update protocol (#152).