API support for multiple views
This is _not_ an actual implementation of multiple views, but rather
handling of the `new_view` RPC to create a new buffer, open a file, or
open a view into an existing buffer, as appropriate.

This also moves the `save` rpc to be at the Tab level; this it easier to
track lifecycle and to associate buffers with files, which we need in
order to support multiple views.