Adopt multi-view API in Editor
The editor now behaves as if it could have multiple views. For the time
being, the editor still has a `view` member, which refers to the
currently active view. In current practice this doesn't change, and when
it _does_ change we will have to reconsider how we handle view updating
more generally.

After this commit, `TabCtx` is sort of an orphan, and how we route
messages around probably needs a rethink.