Keep track of saved/unsaved state
This initial implementation keeps track of a saved revision. This
revision is initially set to the latest revision when creating a new
editor and is updated to the latest revision during a save.

The unsaved state is calculated by comparing the deletion subset of both
revisions. This creates the following behavior: If new text is inserted
and the deleted back to the original state, there are no unsaved
changes; however, if existing text is deleted and retyped, there are now
unsaved changes. If you undo back to the state of the save revision,
there are no unsaved changes.