Implemented receiving and applying plugin edits
This is a preliminary attempt at applying plugin edits. It handles only
simple, single edits, and only edits that occur in response to an
`update` rpc.

As plugins receive updates for edits occuring from themselves and from
other plugins, this introduces the concept of the edit 'authorship'.
Currently, the author is either a view or a plugin, with the plugin
responsible for choosing an identifier; eventually identifiers should be
selected by core.