Do view rendering using new multi-selection
This is part of a migration to the new multi-selection approach. The
old cursor/selection state (which is still what's used in edit
operations) is copied over (when dirty) to the new format. Then, all
rendering is done with the new format.

Most of the logic upstream affinity is in place, but it will require
some small changes. I suggest we get back to that when we can generate
upstream positions from mouse clicks.

Another step towards #188