Improve cursor position handling for plugin edits
When applying plugin edits that are added after the cursor, the
cursor position has to be reset to stay before this update and not
shift to the right.

Changes between the head and plugin edit revision have to be taken
into account to position the cursor correctly to properly reflect the
user’s intention.