Fixed out-of-bounds in Editor::cursor_start
The function was not accounting for the fact that the cursor could be at
the beginning or end of the selection region, and that `cursor_col` was
set based on the cursor position. If the cursor was at the end of the
selection, the subtraction would over(under)flow.

This fix first moves the cursor to the `sel_min`, and then calls the
existing function for moving to the start of the current line.

- fixes #228