Implement soft spans
Soft spans are part of the text that are replaced when overrides with
the same text.

This is useful for plugin inserts, e.g., plugins may auto-pair braces:
{|} (cursor is represented with |) where the closing brace is added as
a soft span. Inserting a closing brace } would in this case move the
cursor over the closing brace } instead of adding another one (i.e.
adding } would lead to {}| ).