Compute HorizPos lazily
This commit changes the logic for HorizPos to compute it on the first
vertical motion after a sequence of non vertical motion changes. The
old logic was eager, and always computed it for cursor changes other
than vertical motion (the cursor_col field wasn't even an option).

The behavior should be the same (modulo some edge cases such as moving
up from the first line; in general, the new behavior should match
Sublime 3). The main benefit is that it should be much easier for
cursor changes to maintain the invariant (in general, they just set
horiz to None).

Progress towards #188