Link to May 2017 roadmap
We've been talking about where development and features fit on a
roadmap for a while, so now seems like a good time to formalize it.
This PR originally contained a roadmap doc, but we have moved it to
an issue (#248) for greater flexibility.

This commit links to that roadmap from the README, and also to provide
more up-to-date information about the status of front-ends.

Many thanks to @cmyr for the first draft!