May 2017 roadmap
We've been talking about where development and features fit on a
roadmap for a while, so now seems like a good time to formalize it.

This commit introduces a roadmap as of May 2017, at least roughly
capturing our thinking. It also updates the main README a bit, to
refer to the roadmap and also provide more up-to-date information
about the status of front-ends.

Many thanks to @cmyr for the first draft!