Rename methods of Subset to be more consistent.
There was previously an inconsistency in the naming in Engine and the naming
in Subset. See #256

The naming of the methods of Subset was as if the Subset represented
the complement of the ranges it contained, which if the ranges were deletions
meant it represented the elements that were kept.

However, this was inconsistent with the naming in Engine which had subsets like
`inserts` and `deletes` that according to the concept of Subset actually represented
the non-inserted and non-deleted ranges.

This fixes the inconsistency by renaming methods as if Subset just represents the
ranges it contains, which eliminates having to reason about the complement when
thinking about the entire system.