Mark insert-only deltas using a wrapper struct.
Some methods of Delta only really applied to ones that only contained
inserts. So rather than rely on the caller to ensure that, we use the
type system to check it. Miraculously due to the power of type
inference and Deref, this didn’t require any changes to uses since they
were all correct.