Make more editing commands multi-select aware
The following editing commands are now based on Selection rather than
the old single cursor:

* deletions
* cut/copy
* transpose
* select all
* insert tab
* do_click

In addition, the cursor fix-up logic in apply_plugin_edit has been
removed for the time being. The correct handling is to fully implement
priority, rather than do an explicit modification of the selection.

There are a number of TODOs remaining, and this patch does not delete
the old selection state, but at this point all editing operations
should be updated not to rely on it.

Progress towards #188