Swap order of transform_shrink arguments to make more sense.
transform_expand, transform_union and Delta::transform_shrink all take the
transform as the parameter and return something with the changes of self.
Before this commit, transform_shrink was different for no apparent reason.

This weirdness has caused me lots of extra thinking and a couple bugs,
this commit fixes it.