Fix broken quickcheck test caused by #309
The argument order of transform_shrink was swapped in #309 for multiset but not
the old subset. This caused the quickcheck tests to fail, but for some reason
Travis didn't mark it as failing on Github. I think I might have gotten an email
but the PR got a green checkmark.

This fixes it by just swapping the test, not the old implementation.