Combine sequential edits with same priority in merge
Applies an optimization where it combines sequential revisions with the
same priority into one transform to decrease the number of transforms that
have to be considered in `rebase` substantially for normal editing
patterns. Any large runs of typing in the same place by the same user (e.g
typing a paragraph) will be combined into a single segment in a transform
as opposed to thousands of revisions.