Refactor revision IDs into RevId struct and u64 tokens
The RevId struct is a delta-compressible representation used by Engine, the
RevToken is an alias for u64 that is used by plugins as a non-colliding
identifier for a revision to base their edits on.

The RevToken is currently just a DefaultHasher hashed version of RevId,
which since DefaultHasher is cryptographically-pseudorandom should never
really collide. The collision set is just concurrent revisions since it
searches for the corresponding revision from newest to oldest.

This refactor paves the way for different Fuchsia devices to have separate
device IDs on their Engine, allowing them to generate non-conflicting
revisions that can be merged using the Ledger syncing mechanics.