Switch to 96 bits of session ID
The numbers with 64 bits were such that if Fuchsia/Xi was a spectacular success
and a million people made documents with 6 million sessions there would be on
expectation one person who's document got corrupted by a collision.

Realistically not enough people would have 6 million sessions on a single
document but I could avoid the possibility entirely without increasing
struct size so why not.