Use trait object for RpcPeer
This change allows us to ignore the peer's concrete type, which has the
practical effect of removing all the places where we previously had to
pass around a generic <W: Write> parameter.

It will hopefully have the secondary effect of allowing for a cleaner
modification of the rpc system to support typed handlers and peers.