RS11 refinement WIP
This commit is for stopping the propagation of state rippling but
is not done yet.

Checkpoint: update applies a finer-grained delta to state cache.

Checkpoint: tries to test equality, but that's failing.

Partial resolution: own syntect git branch containing fix for
https://github.com/trishume/syntect/issues/105 . Will need to point
back to crates.io syntect once that's released.

Checkpoint: works, but frontier not perfect.

Checkpoint: logic for updating frontier should be working now.