Rpc run loop now returns a Result<(), ReadError>
This is primarily to facilitate the writing of certain types of tests,
specifically those where we want to start a run loop with a specific
handler, and then verify that JSON is handled as expected.

This also has the benefit of providing more error information to the
call site, where more specific decisions can be made about error
handling and/or logging.