Protocol support for getting/modifying user config
This adds two methods to the RPC protocol, and removes the
`DebugOverrideSetting` method that was used during development.

'modify_user_config { domain, table }' is the new method for updating or
deleting entries in a given config domain.

'get_config { view_id }' is the new method for fetching the collated
config table applicable to a particular view.