Start implementation of minimal invalidation
This patch uses a fancy data structure to track the state of the
front-end's line cache on a fine grained basis, and construct a
minimal delta to be sent on update.

As of this patch, the data structure is done, but most changes to
the view just invalidate the whole thing. Thus, the visible effect
is minimal. That said, the core is in control of eviction of the
front-end's cache, so the unbounded growth is not possible, and it
also won't keep resending the same content, which should help with
scrolling performance.

WIP (not done)

Fixes #270 and is good progress towards #317.