Reload files when changed on disk
This uses modification times to check if a file has been modified by
another process. If an open file with no unsaved changes is modified on
disk, the changed file is reloaded silently; If a file changes on disk
and _does_ have unsaved changes, we pop an alert if the user tries to
save.

This last behaviour is definitely not ideal, but seemed like an okay
first pass.

closes #362