Ensure core shuts down after an error
Previously if an error occured in the main thread, xi-core would hang
indefinitely because it is waiting on the read thread, which doesn't
know about the error.

With this change, we keep track of crashes via an AtomicBool, which
the read thread checks before each new read operation.

Because reads are blocking, core will not exit if an error occurs
_while_ thre read thread is blocked, until the next RPC is received.
This isn't ideal, but is a limitation of blocking I/O.

Fixes #434