Add experimental width-based wrapping
This change adds a width measurement RPC, and uses it to compute
breaks based on width.

We also add a cache for width measurements, with an interface
supporting aggregating a bunch of measurements together into a single
batch RPC.

A number of things not ideal yet. Incremental rewrapping is not width
based. The trailing space is considered part of the width.  Only one
request batch is used no matter how large the document, so it won't
scale well.