Refactored indent and outdent to reuse code, changed outdent logic
changed outdent to try and keep cursor position while not deleting lines that have a space which immediately follows.
previously, outdent checked up to the tab size only and deletes if the slice up to the tab size contains a whitespace.
outdent now tries to find the first non-whitespace character if the line is not fully indented (less than tab size)