Put back multi-plugin support
This is a quick temporary (famous last words) fix to allow support for
running multiple plugins. This is chiefly intended to enable the GSoC
LSP work.

With this patch, all plugins are launched on startup, and are global to
all views. Plugins can be manually stopped and started via the debug
menu.