Cleanup core init and ConfigManager
A long time ago we had a constraint where we didn't know config dirs
when creating CoreState (then Tabs), and so we had to do that setup
afterwards.

That constraint has gone away, but the code hadn't caught up. This gets
us most of the way there.

In addition, it includes a bit of API simplification in `ConfigManager`.
These changes are in anticipation of some upcoming work that will make
plugins responsible for defining new languages/syntaxes.