ConfigManager tracks all config state
This is another batch of cleanup & refactoring. In pointform:

- the language assigned to a buffer is tracked in the ConfigManager
- the buffer's config is tracked in the ConfigManager. This means that
we don't need to have a config in order to init an `Editor`, which means
we now only have one code path for creating a new view.
- configs are passed into `Editor` and `View` as needed; this means the
`View` doesn't need to reach in to the `Editor` to access the config.

This will enable and/or simplify some additional feature work:
- allowing the client to set a view's language without saving
- implementing width wrapping via the config system
- auto launching plugins based on language (and other) tags