Add core->client new_view RPC
This adds a core->client RPC for opening a new view. It also adds a
client->core RPC, `new_view_async`, that is mostly intended to test that
thew new core RPC works. I'm not sure if it should stick around or not.
:|

Regardless this needs docs before it should be merged.