Fix incorrectly trying to fetch implicit last line
self.num_lines is a count, and line_num is an index; if line_num ==
num_lines, it means line_num is the 'implicit' last line, which doesn't
actually exist.

I haven't thought about this too much; there's possibly an argument for
returning the empty string here, but I'm not sure that's right.

I suspect that this fixes #721.