Add Edit type (alternative delta representation)
This is tentative work, motivated by wanting to improve plugin
ergonomics. The idea is that a delta can be cheaply converted into
a sequence of edits, which more intuitively represent the changes to a
document.

Problems with this, as currently written:

- the implementation is a bit gross, although this is fixable.
- There's a lot of overlap with the 'iter_inserts' and 'iter_deletions'
methods.
- currently the 'Edit' struct only makes sense in the coordinate space
of the pre-edit document, although in the plugin-lib we only notify the
plugin _after_ applying the delta, so a sequence of edits is not going
to be super meaningful. This is an interesting design problem. It would
not be difficult to include a 'delta_position' value with an edit that
could be used to determine the start/end in the coordinate space of the
post-edit document.