Update copyright messages to xi-editor
Following guidance from https://opensource.google.com/docs/releasing/authors/
this patch changes the copyright message to "The xi-editor Authors" and
creates an AUTHORS file for contributors to record their name.

Also minor update to the contribution guidelines to remind people to add their
name to AUTHORS, and tweak the rules for approving and merging patches.