Compute deltas from arbitrary ropes
This adds a preliminary algorithm for computing an efficient delta
between two ropes.

This algorithm produces poor deltas in a few pathological cases, such as
when a document is compared against itself with its lines shuffled, or
when a pair of documents contains few line breaks and have numerous small
edits; however it has consistently quick running time.