Fix injection of view_id into EditCommands
We are injecting view_id into the params section of an EditNotification.
However the resulting request is incorrect and cannot be correctly
deserialized.

We are now correctly injecting it into the top level of the
serialized EditNotification. We are adding a simple test that ensure
this invariant stays intact. This change makes it easier to use
xi_core_lib::rpc as a crate to construct a valid request.