Improve sse4.2 string compare performance
This version just uses bytewise compare, instead of the fancier (and
slower) string-specific operations. This shows a marked improvement, and
the new sse version is actually faster than the avx version, which is
slightly surprising; it may be that the avx loads are higher latency and
that's dominating? I've also only tested on one machine, a laptop with
turbo-boost disabled.