Add IntervalBounds trait and audit use of Interval
This adds the IntervalBounds trait, a simplification on the stdlib
RangeBounds trait. This allows the full set of range notations (0..,
0..=6, ..6, .., 2..4, 2..=5) to be used in places that accept intervals,
where appropriate.

Where it is _not_ appropriate, the bounded range types also implement
Into<Interval>, so they can be made available on their own.

This patch also includes changes to all public use of `Interval` to take
advantage of these new conversions.