Fix incorrect invalidation of breaks on rewrap
This is a quick fix to the current width wrapping implementation,
patching over the major issue where an edit (delta) that ended on a byte
preceding a newline (and maybe also a soft break?) would cause the rest
of the document's breaks to be incorrectly invalidated.

This fix is likely papering over some fundamental problems with the
BreaksMetric, but it does at least fix the behaviour, and is fairly
cheap; the only consequence is that we will in some instances
recalculate an additional line's worth of breaks after some edits.

I intend to look at the breaks data structures more closely shortly, but
this seems worth a quick patch.

 ¯\_(ツ)_/¯

fixes https://github.com/xi-editor/xi-mac/issues/301