Add surrounding_pairs option to xi-core-lib; add InsertDrift
This adds the surrounding_pairs option. If all selections are regions
(and not carets) and the user types a first half of a surrounding pair,
then we will surround the selection with the surrounding pair instead
of replacing the selection. This config option can be edited on a per-
language basis.

This change also adds InsertDrift as a replacement for the old
keep_selections option. The keep_selections bool, if true, overrode
the 'after' property, ensuring that inserts 'drifted' to the side of
a cursor that was inside a selection. This change keeps that same
functionality, but adds the option to have inserts 'drift' to the
outside of a selection instead.