Implement comment toggling in the syntect plugin
This adds syntax-aware comment toggling.

The implementation is sort of weird, in a way that maybe
highlights some of the general weirdness around this feature,
and also maybe illustrates some future directions for plugin-
provided commands.

Basically: comment toggling, like auto-indent, is syntax
dependant. Unlike auto-indent, however, it is also the result
of an explicit user action.

This is a funny case. It seems silly to add a 'toggle_comment'
command to the RPC protocol. We do need to route through core,
because we need to extract the lines to be commented; plugins
do not currently have access to selection state. However we only
expect this to be handled by one plugin.

So: this is a hack. This has given me some ideas about how
plugins can declare their capabilities, though; I'll open an
issue for that in a little bit.

In the meantime, this is a useful feature.