Unify width and column wrapping
This reimplements column-based wrapping using the existing width-based
system, which reduces the number of code paths for both creating and updating
breaks by up to 50%!

This is currently _slightly_ hacky; we have to clear the width cache when
switching between word and column wrapping, because column wrapping now
uses the width cache but stores codepoint counts instead of actual
measured widths.

This is temporary; assuming we keep column-based wrapping around at all,
it will make sense in the future to get the 'unit width' for a given
style from the frontend, and use that + UAX#11 to generate actual
widths.