Do minimal invalidation with soft breaks enabled
This is a little complicated, because we need different information
after an edit than we do when we're just rewrapping. Most importantly,
when an edit occurs it's possible that the edit deletes existing
breaks; this means we need to count the breaks in the pre-edit
rope and breaks structures in order to handle invalidation correctly.