Rethink boundaries on can_fragment metrics
After some thought and discussion, we've decided that there should not
automatically be boundaries at the beginning and end for metrics that
are not can_fragment. This change removes those from rope.

It's expected that it'll break things that depend on the old behavior,
but we're splitting that out.

Also contains some documentation improvements.