Keep viewport anchored during incremental rewrap
This introduces the concept of anchors and drift. Well, shift. I was
calling it drift in my head but it's shift in the code, which is less
poetic but more clearly suggests intentionality.

To state the problem: when we're calculating new line breaks, we do it
incrementally; that is, we calculate the breaks for the visible region,
render that immediately, and then incrementally calculate the breaks for
the rest of the document. The problem here comes when the incremental
work is happening above the visible region; this can change the number
of visual lines in the document, which means that the line the client
thinks is line 50 is now line 68; this means the view will jump
unexpectedly as these updates come in.

The solution is to pick an anchor point during rewrapping; in this
commit the anchor point is the offset of the newline character preceding
the first visible line.

We wrap the lines between this anchor point and the end of the view
immediately; all subsequent batches of wrapping happen either above or
below this point. If it happens above, we keep a running tally of the
number of lines added or removed; during the next update, we take this
number (a signed integer) and send it to the frontend as the "shift"
field of the "update" RPC. the frontend then reorients itself by
translating what it thinks its current position is by the "shift" value.
This way the view stays fixed, even as relative line numbers change.