Fix linewrap iterator to handle invalid text iterator
The doc for `Cursor::next` promises that when no next boundary is found,
the cursor becomes invalid (meaning that leaf will be `None`). However,
the old implementation often (but not always!) left the leaf valid even
at the end.

The implementation for the `MergedBreaks` iterator in linewrap depended
on this behavior, requesting the leaf to check for trailing newline.

I suspect the logic can be simplified, as it can check whether the leaf
is none, rather than testing for newline explicitly. But that seems like
a followup, rather than in the scope of this patch. I just added a call
to set the cursor (ensuring it's valid) before testing the leaf.

This patch adds a test ensuring that leaf is invalidated as promised (a
test the old implementation fails).