Only check for x86 features on x86_64
This fixes an issue where when compiled for x86 (i686) we might
try to call a function that had not actually been compiled.

This brings up an obvious question, which is whether we should be
using at avx2/sse4.2 on i686? My understanding was that no i686
processors existed with avx2 support, and only a couple with sse4.2,
and so this didn't really feel worth it at the time.

- fixes #1097