[xi-unicode] Explicit breaks at end of text
Instead of always indicating end of text as a hard break, report whether
the text ends with an explicit line break.

Document the behavior.

Document the line break properties, and also make the ordering
consistent with ICU.

This is a breaking change but hopefully should be useful and not cause
too much trouble.

Fixes #1289